requirements.txt
Flask>=2.0
biopython>=1.79